2024-2028/Kowshik/Semantic-Segmentation-on-Cityscapes-using-U-Net-MobileNetV2-main/requirements.txt
tensorflow
keras
numpy
opencv-python
matplotlib
Pillow
scikit-learn
